requirements.txt
pandas
nltk
matplotlib
transformers
gensim
torch
scikit-learn
openai
langchain
llama-index
python-dotenv
pinecone-client